+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
+- Added support for InvokableResolver, ReflectionResolver and StaticFactoryResolver to dismiss the service key for a
+  shorter config notation. The service value will be used as the key.
+
 ### Changed
 
 ### Deprecated